moved generic functions to utils
